Game now creates a Config File in Documents Folder

+++ src/org/engine/logic/CustomMath.java
@@ -3,7 +3,7 @@ package org.engine.logic;
 public class CustomMath
 {
     // Constricts a Value to a Min and a Max
-    public static float clamp(float var, float min, float max)
+    public static double clamp(double var, double min, double max)
     {
         if(var>= max)
             return var = max;
@@ -14,13 +14,13 @@ public class CustomMath
     }
 
     // Smoothly Move one Value to Another
-    public static float lerp(float a, float b, float t)
+    public static double lerp(double a, double b, double t)
     {
         return a + t * (b - a);
     }
 
-    public static float moveToward(float a, float b, float t)
+    public static double moveToward(double a, double b, double t)
     {
-        return (float)(a + Math.signum(b - a) * t);
+        return (a + Math.signum(b - a) * t);
     }
 }
